# [5.7] Use the time function for test performance