Benchmark: 
```php
<?php

$microtime = microtime(true);

for ($i = 1; $i < 5000000; ++$i) {
    time();
}

var_dump(microtime(true) - $microtime);


$microtime = microtime(true);

for ($i = 1; $i < 5000000; ++$i) {
    new DateTime();
}

var_dump(microtime(true) - $microtime);
```

Results
```
new DateTime() (50000x): 0.62s
time() (50000x): 0.00s

new DateTime() (500000x): 5.89s
time() (500000x): 0.04s

new DateTime() (5000000x): 62.66s
time() (5000000x): 0.45s
```